requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
biopython==1.85
cachetools==5.5.2
cafe==2.0.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
Django==5.2
eli5==0.10.1
fonttools==4.58.0
frozenlist==1.6.0
graphviz==0.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imbalanced-learn==0.6.1
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.2.3
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests_ntlm==1.3.0
scikit-learn==1.6.1
scikit-plot==0.3.7
scipy==1.15.3
seaborn==0.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sspilib==0.3.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wfastcgi==3.0.0
whitenoise==6.9.0
yarl==1.20.0
